% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtong_functions.R
\name{subst}
\alias{subst}
\title{Whenever the 'string' encounters @ symbol it will be replaced with 'replacement'; 'replacement' can be a vector}
\usage{
subst(string, replacement, collapse = ", ")
}
\description{
Whenever the 'string' encounters @ symbol it will be replaced with 'replacement'; 'replacement' can be a vector
}
